Move to Basic::Bot and rename Package to be conform
Removed the old deprecated Net::IRC Module and moved to Basic::Bot
there is a lot of work to do to work properly with Basic::Bot but the most
is done.

Renamed and rearranged the whole Package Tree to be more conform
to other Perl Modules and to be more easy to read.
A lot of Sub Modules had to be rewritten or fixed due to object oriented
issues that are not relevant at the moment.